requirements.txt
# Core ML and AI libraries
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.24.0
openai>=1.0.0

# Web framework
streamlit>=1.28.0

# Data processing
nltk>=3.8
beautifulsoup4>=4.12.0
python-dateutil>=2.8.0

# Environment management
python-dotenv>=1.0.0

# Email processing (built-in modules)
# mailbox - built-in Python module
# email - built-in Python module
